refactor: remove theft reporting
