+++ vault/tests/vault_integration_tests.rs
@@ -82,250 +82,6 @@ where
 
     execute(client, vault_id, vault_provider).await
 }
-// seed is to make txid unique for each payment
-async fn pay_redeem_from_vault_wallet(
-    vault_provider: InterBtcParachain,
-    btc_rpc: MockBitcoinCore,
-    addr_seed: u8,
-    vault_id: VaultId,
-) {
-    // Theft detection works by extracting the public address from transaction inputs,
-    // and comparing them to registered vault addresses.
-    // We can't easily create a spending input corresponding to a output address, because
-    // the input script does not contain the public key; rather, it contains a hash `x`
-    // such that `hash(x)` is equal to the output address.
-    // To work around this, we create an arbitrary spending script, and then calculate the
-    // corresponding output address. Then, we make the vault register the address and we
-    // proceed to spend it.
-
-    let btc_rpc = &btc_rpc;
-    let vault_provider = &vault_provider;
-    let vault_id = &vault_id;
-    vault_provider
-        .on_event::<RequestRedeemEvent, _, _, _>(
-            |event| async move {
-                tracing::error!("Event {}", addr_seed);
-                let request = vault_provider.get_redeem_request(event.redeem_id).await.unwrap();
-                // step 1: create a spending transaction from some arbitrary address
-                let mut transaction = btc_rpc
-                    .create_transaction(
-                        request.btc_address,
-                        request.amount_btc as u64,
-                        1000,
-                        Some(event.redeem_id),
-                    )
-                    .await
-                    .unwrap();
-                let mut x = [3; 33];
-                x[1] = addr_seed;
-                // set the hash in the input script. Note: p2wpkh needs to start with 2 or 3
-                transaction.transaction.input[0].witness = vec![vec![], x.to_vec()];
-                // make txid unique
-                transaction.transaction.input[0].previous_output.vout = addr_seed as u32;
-                // first byte of the script needs to be non-zero in order to be parsed as p2wpkh. If we don't overwrite
-                // this, the client will parse the input as p2wpkh, while the parachain parses is as a script hash
-                // address
-                transaction.transaction.input[0].script_sig = bitcoin::Script::from(vec![
-                    1, 71, 48, 68, 2, 32, 91, 128, 41, 150, 96, 53, 187, 63, 230, 129, 53, 234, 210, 186, 21, 187, 98,
-                    38, 255, 112, 30, 27, 228, 29, 132, 140, 155, 62, 123, 216, 232, 168, 2, 32, 72, 126, 179, 207,
-                    142, 8, 99, 8, 32, 78, 244, 166, 106, 160, 207, 227, 61, 210, 172, 234, 234, 93, 59, 159, 79, 12,
-                    194, 240, 212, 3, 120, 50, 1, 71, 81, 33, 3, 113, 209, 131, 177, 9, 29, 242, 229, 15, 217, 247,
-                    165, 78, 111, 80, 79, 50, 200, 117, 80, 30, 233, 210, 167, 133, 175, 62, 253, 134, 127, 212, 51,
-                    33, 2, 128, 200, 184, 235, 148, 25, 43, 34, 28, 173, 55, 54, 189, 164, 187, 243, 243, 152, 7, 84,
-                    210, 85, 156, 238, 77, 97, 188, 240, 162, 197, 105, 62, 82, 174,
-                ]);
-
-                // extract the public address corresponding to the input script
-                let input_address = transaction.transaction.extract_input_addresses::<BtcAddress>()[0];
-                tracing::error!("txid {} {}", addr_seed, transaction.transaction.txid());
-                // now make the vault register it
-                assert_ok!(vault_provider.register_address(vault_id, input_address).await);
-                tracing::error!("Registered {}", addr_seed);
-                let return_to_self_address = transaction.transaction.extract_output_addresses::<BtcAddress>()[1];
-                // register return-to-self address if it hasnt been yet
-                let wallet = vault_provider.get_vault(vault_id).await.unwrap().wallet;
-                if !wallet.addresses.contains(&return_to_self_address) {
-                    vault_provider
-                        .register_address(vault_id, return_to_self_address)
-                        .await
-                        .unwrap();
-                }
-
-                // now perform the theft
-                assert_ok!(btc_rpc.send_transaction(transaction).await);
-                tracing::error!("sent {}", addr_seed);
-            },
-            |_err| (),
-        )
-        .await
-        .unwrap();
-}
-
-#[tokio::test(flavor = "multi_thread")]
-async fn test_report_vault_theft_succeeds() {
-    service::init_subscriber();
-
-    let (client, _tmp_dir) = default_provider_client(AccountKeyring::Alice).await;
-
-    let root_provider = setup_provider(client.clone(), AccountKeyring::Alice).await;
-    let relayer_provider = setup_provider(client.clone(), AccountKeyring::Bob).await;
-    let vault_provider = setup_provider(client.clone(), AccountKeyring::Charlie).await;
-    let vault_id = VaultId::new(
-        AccountKeyring::Charlie.into(),
-        DEFAULT_TESTING_CURRENCY,
-        DEFAULT_WRAPPED_CURRENCY,
-    );
-
-    set_exchange_rate_and_wait(
-        &relayer_provider,
-        DEFAULT_TESTING_CURRENCY,
-        FixedU128::saturating_from_rational(1u128, 100u128),
-    )
-    .await;
-
-    assert_ok!(
-        try_join(
-            root_provider.set_bitcoin_confirmations(0),
-            root_provider.set_parachain_confirmations(0),
-        )
-        .await
-    );
-
-    let btc_rpc = MockBitcoinCore::new(relayer_provider.clone()).await;
-
-    let issue_amount = 100000;
-    let vault_collateral =
-        get_required_vault_collateral_for_issue(&vault_provider, issue_amount, vault_id.collateral_currency()).await;
-    assert_ok!(
-        vault_provider
-            .register_vault_with_public_key(&vault_id, vault_collateral, btc_rpc.get_new_public_key().await.unwrap(),)
-            .await
-    );
-
-    let vaults = Arc::new(vault::Vaults::from(Default::default()));
-    let random_delay = ZeroDelay;
-
-    test_service(
-        join(
-            vault::service::monitor_btc_txs(
-                btc_rpc.clone(),
-                relayer_provider.clone(),
-                random_delay.clone(),
-                0,
-                vaults.clone(),
-            ),
-            vault::service::listen_for_wallet_updates(relayer_provider.clone(), btc_rpc.network(), vaults.clone()),
-        ),
-        async {
-            // Theft detection works by extracting the public address from transaction inputs,
-            // and comparing them to registered vault addresses.
-            // We can't easily create a spending input corresponding to a output address, because
-            // the input script does not contain the public key; rather, it contains a hash `x`
-            // such that `hash(x)` is equal to the output address.
-            // To work around this, we create an arbitrary spending script, and then calculate the
-            // corresponding output address. Then, we make the vault register the address and we
-            // proceed to spend it.
-
-            // step 1: create a spending transaction from some arbitrary address
-            let mut transaction = btc_rpc
-                .create_transaction(BtcAddress::P2PKH(H160::from_slice(&[4; 20])), 1500, 1000, None)
-                .await
-                .unwrap();
-            // set the hash in the input script. Note: p2wpkh needs to start with 2 or 3
-            transaction.transaction.input[0].witness = vec![vec![], vec![3; 33]];
-            // first byte of the script needs to be non-zero in order to be parsed as p2wpkh. If we don't overwrite
-            // this, the client will parse the input as p2wpkh, while the parachain parses is as a script hash address
-            transaction.transaction.input[0].script_sig = bitcoin::Script::from(vec![
-                1, 71, 48, 68, 2, 32, 91, 128, 41, 150, 96, 53, 187, 63, 230, 129, 53, 234, 210, 186, 21, 187, 98, 38,
-                255, 112, 30, 27, 228, 29, 132, 140, 155, 62, 123, 216, 232, 168, 2, 32, 72, 126, 179, 207, 142, 8, 99,
-                8, 32, 78, 244, 166, 106, 160, 207, 227, 61, 210, 172, 234, 234, 93, 59, 159, 79, 12, 194, 240, 212, 3,
-                120, 50, 1, 71, 81, 33, 3, 113, 209, 131, 177, 9, 29, 242, 229, 15, 217, 247, 165, 78, 111, 80, 79, 50,
-                200, 117, 80, 30, 233, 210, 167, 133, 175, 62, 253, 134, 127, 212, 51, 33, 2, 128, 200, 184, 235, 148,
-                25, 43, 34, 28, 173, 55, 54, 189, 164, 187, 243, 243, 152, 7, 84, 210, 85, 156, 238, 77, 97, 188, 240,
-                162, 197, 105, 62, 82, 174,
-            ]);
-
-            // extract the public address corresponding to the input script
-            let input_address = transaction.transaction.extract_input_addresses::<BtcAddress>()[0];
-            // now make the vault register it
-            assert_ok!(vault_provider.register_address(&vault_id, input_address).await);
-
-            // now perform the theft
-            assert_ok!(btc_rpc.send_transaction(transaction).await);
-
-            assert_event::<VaultTheftEvent, _>(TIMEOUT, vault_provider, |_| true).await;
-        },
-    )
-    .await
-}
-
-#[tokio::test(flavor = "multi_thread")]
-async fn test_report_vault_double_payment_succeeds() {
-    test_with_vault(|client, vault_id, vault_provider| async move {
-        let root_provider = setup_provider(client.clone(), AccountKeyring::Alice).await;
-        let relayer_provider = setup_provider(client.clone(), AccountKeyring::Bob).await;
-        let user_provider = setup_provider(client.clone(), AccountKeyring::Dave).await;
-
-        // set_exchange_rate(&relayer_provider, FixedU128::saturating_from_rational(1u128, 100u128)).await;
-
-        assert_ok!(
-            try_join(
-                root_provider.set_bitcoin_confirmations(0),
-                root_provider.set_parachain_confirmations(0),
-            )
-            .await
-        );
-
-        let btc_rpc = MockBitcoinCore::new(relayer_provider.clone()).await;
-
-        let issue_amount = 100000;
-        let vault_collateral =
-            get_required_vault_collateral_for_issue(&vault_provider, issue_amount, vault_id.collateral_currency())
-                .await;
-        assert_ok!(
-            vault_provider
-                .register_vault_with_public_key(
-                    &vault_id,
-                    vault_collateral,
-                    btc_rpc.get_new_public_key().await.unwrap(),
-                )
-                .await
-        );
-        assert_issue(&user_provider, &btc_rpc, &vault_id, issue_amount).await;
-
-        let vaults = Arc::new(vault::Vaults::from(Default::default()));
-
-        let random_delay = ZeroDelay;
-
-        // we make the vault start two listen_for_redeem_requests processes, this way there will be a double payment
-        // that should be reported
-        test_service(
-            join4(
-                vault::service::monitor_btc_txs(
-                    btc_rpc.clone(),
-                    relayer_provider.clone(),
-                    random_delay.clone(),
-                    0,
-                    vaults.clone(),
-                ),
-                vault::service::listen_for_wallet_updates(relayer_provider.clone(), btc_rpc.network(), vaults.clone()),
-                pay_redeem_from_vault_wallet(vault_provider.clone(), btc_rpc.clone(), 2, vault_id.clone()),
-                pay_redeem_from_vault_wallet(vault_provider.clone(), btc_rpc.clone(), 3, vault_id.clone()),
-            ),
-            async {
-                let address = BtcAddress::P2PKH(H160::from_slice(&[2; 20]));
-                user_provider
-                    .request_redeem(issue_amount / 2, address, &vault_id)
-                    .await
-                    .unwrap();
-
-                assert_event::<VaultDoublePaymentEvent, _>(Duration::from_secs(120), root_provider, |_| true).await;
-            },
-        )
-        .await;
-    })
-    .await;
-}
 
 #[tokio::test(flavor = "multi_thread")]
 async fn test_redeem_succeeds() {
@@ -867,13 +623,6 @@ async fn test_refund_succeeds() {
             // make the vault register the input used in that transaction
             let input_address = tx.extract_input_addresses::<BtcAddress>()[0];
             assert_ok!(vault_provider.register_address(&vault_id, input_address).await);
-
-            // check that it is not seen as theft
-            let metadata = btc_rpc.wait_for_transaction_metadata(tx.txid(), 0).await.unwrap();
-            let result = user_provider
-                .report_vault_theft(&vault_id, &metadata.proof, &metadata.raw_tx)
-                .await;
-            assert!(result.unwrap_err().is_valid_refund());
         };
 
         test_service(refund_service, fut_user).await;
